backend-whatsapp-webhook/requirements.txt
functions-framework==3.*
flask==3.0.0
google-cloud-firestore>=2.11.0
google-cloud-aiplatform>=1.38.0
